.Exerd 무결성위한 관계개선 MAX(NUM) NULL 리턴(500 Error)이 문제없도록 선조치 (리턴타입 int -> Object로)

+++ web03t/src/main/java/io/helloworld/service/ItemService.java
@@ -103,7 +103,7 @@ public class ItemService {
 	public int getTourDayCount(int itemNo) {
 		int count = 1;
 		if ((Integer)itemDao.selectTourDayCount(itemNo) != null) 
-		count = itemDao.selectTourDayCount(itemNo);
+		count = (Integer)itemDao.selectTourDayCount(itemNo);
 		return count;
 	}
